# api/openapi.yaml
openapi: 3.0.3
info:
  title: Loom API
  description: |
    The Loom API provides programmatic access to the agent orchestration system.
    
    ## Authentication
    
    API requests must include an API key in the `X-API-Key` header when authentication is enabled.
    
    ## Security
    
    - All production deployments should use HTTPS
    - PKI/TLS certificates can be configured for mutual TLS authentication
    - API keys should be rotated regularly
    
  version: 1.0.0
  contact:
    name: Loom Support
    url: https://github.com/jordanhubbard/loom
servers:
  - url: https://localhost:8443
    description: HTTPS server (production)
  - url: http://localhost:8080
    description: HTTP server (development only)
security:
  - ApiKeyAuth: []
paths:
...
  /api/v1/agents:
    get:
      summary: List all agents
      description: Get a list of all running agents
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Agent'

    post:
      summary: Create a new agent
      description: Spawn a new agent with a specified persona
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - persona_name
                - project_id
              properties:
                name:
                  type: string
                  description: Optional custom name for the agent
                persona_name:
                  type: string
                  description: Name of the persona to use
                project_id:
                  type: string
                  description: ID of the project to work on
      responses:
        '201':
          description: Agent created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          description: Invalid request
...
  /api/v1/projects:
    get:
      summary: List all projects
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'

    post:
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Agent:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        persona_name:
          type: string
        status:
          type: string
          enum: [idle, working, deciding, blocked]
        current_bead:
          type: string
        project_id:
          type: string
        started_at:
          type: string
          format: date-time
        last_active:
          type: string
          format: date-time
    Project:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        git_repo:
          type: string
        branch:
          type: string
        beads_path:
          type: string
        context:
          type: object
          additionalProperties:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        agents:
          type: array
          items:
            type: string
    ProjectCreate:
      type: object
      required:
        - name
        - git_repo
        - branch
      properties:
        name:
          type: string
        git_repo:
          type: string
        branch:
          type: string
        beads_path:
          type: string
          default: .beads
        context:
          type: object
          additionalProperties:
            type: string
...